SkyXplore - Decreasing Satiety

+++ apphub-lib/apphub-lib-common_util/src/main/java/com/github/saphyra/apphub/lib/common_util/SleepService.java
@@ -1,7 +1,7 @@
 package com.github.saphyra.apphub.lib.common_util;
 
 public class SleepService {
-    public void sleep(int millis) {
+    public void sleep(long millis) {
         try {
             Thread.sleep(millis);
         } catch (InterruptedException e) {
